+++ packages/sst/package.json
@@ -26,7 +26,7 @@
     "./constructs/future": "./constructs/future/index.js",
     "./constructs": "./constructs/index.js",
     "./context": "./context/index.js",
-    "./node/future/*": "./node/*/index.js",
+    "./node/future/*": "./node/future/*/index.js",
     "./node/*": "./node/*/index.js",
     ".": "./index.js",
     "./*": "./*"
